Refactor chatbot and server structure, add logging, and implement chatbot router
