+++ cyber_query_ai/routers/__init__.py
@@ -1,0 +1,5 @@
+"""Routers for the CyberQueryAI API."""
+
+from .chatbot_router import ChatbotRouter
+
+__all__ = ["ChatbotRouter"]
